refactor(nef): place scanner tests with the code they exercise
Distribute the tests to the module each drives: the ref-discovery and
import tests to analyze.rs, the ScanError message tests to error.rs, and
the scan_package pipeline tests to mod.rs beside the public API they
cover.
